Generics.Zipper

Description

The main module of the generic-zipper library.

Example

Let us consider the example of a family of mutually recursive datatypes.

 data Expr = Const Int
           | Add Expr Expr
           | Mul Expr Expr
           | EVar Var
           | Let Decl Expr
 data Decl = Var := Expr
           | Seq Decl Decl
 type Var  = String

To use the zipper interface for these datatypes, you need to create instances of the Generic class from the generics-sop library. It is possible to derive the instances via GHC.Generics:

{-# LANGUAGE DeriveGeneric #-}

import Generics.SOP
import qualified GHC.Generics as GHC (Generic)

data Expr = Const Int
          | Add Expr Expr
          | Mul Expr Expr
          | EVar Var
          | Let Decl Expr
    deriving (GHC.Generic, Show)
data Decl = Var := Expr
          | Seq Decl Decl
    deriving (GHC.Generic, Show)
type Var  = String

instance Generic Expr
instance Generic Decl

Assume we want to traverse the following example expression using the zipper.

example :: Expr
example = Let ("x" := Mul (Const 6) (Const 9))
              (Add (EVar "x") (EVar "y"))

Let us define a class that provides a function for updating the tree.

class Update a where
    solve :: a -> a
    solve = id
instance Update Expr where
    solve _ = Const 42
instance Update Decl
instance Update Var

Using the Zipper

A type for a family of mutually recursive datatypes is a promoted list of types that you consider as mutually recursive:

  type ExampleFam1 = '[Expr, Decl, Var]

The flipped function composition >>> and flipped Kleisli (monadic) composition >=> can be used to chain moves and edit operations. In the following definition of test1, enter starts navigation within the tree-like structure of the expression; the enter function takes two type arguments: a family of datatypes and a (single or complex) constraint that is meant to be applied to each datatype in the family (such as Update in our example).

test1 = enter @ExampleFam1 @Update
            >>> goDown >=> goDown >=> goRight >=> update solve
            >>> leave >>> return $ example

The code above uses the TypeApplications extension.

Defining another example of traversal with this value shows that usage of the generic zipper is meant to be flexible for various families of mutually recursive datatypes.

type ExampleFam2 = '[Expr, Var]

test2 = enter @ExampleFam2 @Update
            >>> goDown >=> goDown >=> goRight >=> update solve
            >>> leave >>> return $ example

The calls of these two tests yield the following results:

>>> test1
Just (Let ("x" := Const 42) (Add (EVar "x") (EVar "y")))
>>> test2
Just (Let ("x" := Mul (Const 6) (Const 9)) (Add (EVar "x") (Const 42)))

Locations

data Loc (r :: *) (fam :: [*]) (c :: * -> Constraint)

A location contains the current focus and its context. Its type fixes the type of the root, the type of the family of datatypes, and the constraint over datatypes in the family.

Interface

Movement functions

goDown :: Loc a fam c -> Maybe (Loc a fam c)

Move down to the leftmost child if possible. For leaves, return Nothing.

goRight :: Loc a fam c -> Maybe (Loc a fam c)

Move down to the right sibling if possible, else return Nothing.

goUp :: Loc a fam c -> Maybe (Loc a fam c)

Move down to the parent if possible. For the root, return Nothing.

Starting navigation

enter :: forall fam c a. (Generic a, In a fam, Zipper a fam c) => a -> Loc a fam c

Enter a tree. Places the value into the empty context.

Ending navigation

leave :: Loc a fam c -> a

Move up to the root and return the expression in focus.

Updating

update :: (forall b. c b => b -> b) -> Loc a fam c -> Loc a fam c

Update the current focus, which contains a constrained value.

Combining operations

(>>>) :: (a -> b) -> (b -> c) -> a -> c

Flipped function composition.

Re-exports

(>=>) :: Monad m => (a -> m b) -> (b -> m c) -> a -> m c infixr 1

Left-to-right Kleisli composition of monads.